# Use CRT implementation of memset/memcpy