Few years ago, we already switched to using platform memset/memcpy in linux/arm64 in https://github.com/dotnet/coreclr/pull/17536 and likewise for windows and linux on x64 in https://github.com/dotnet/coreclr/pull/25750. The only left out was windows/arm64 which was still using the hand written assembly code. The platform implementation is highly optimized and has ways to use appropriate routines based on the size of block.

Contributes to https://github.com/dotnet/runtime/issues/67326