@kunalspathak 2022-04-09
```
| Faster                                      | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| StoreBlock.AnyLocation.CopyBlock128         |      1.34 |            11.81 |             8.79 |         |
| StoreBlock.AnyLocation.InitBlockAllZeros128 |      1.33 |             9.48 |             7.12 |         |
| StoreBlock.AnyLocation.InitBlockAllOnes128  |      1.33 |             9.46 |             7.14 |         |
```